# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Bracket Golf API - U.S. Amateur Bracket Challenge
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
}

export interface User {
  id: string;
  name: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export type TournamentStatus = typeof TournamentStatus[keyof typeof TournamentStatus];


export const TournamentStatus = {
  upcoming: 'upcoming',
  active: 'active',
  locked: 'locked',
  completed: 'completed',
} as const;

export interface Tournament {
  id: string;
  name: string;
  year: number;
  slug: string;
  status: TournamentStatus;
  /** @nullable */
  lockTime?: string | null;
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  lastSyncedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type TournamentUpdateStatus = typeof TournamentUpdateStatus[keyof typeof TournamentUpdateStatus];


export const TournamentUpdateStatus = {
  upcoming: 'upcoming',
  active: 'active',
  locked: 'locked',
  completed: 'completed',
} as const;

export interface TournamentUpdate {
  /** @nullable */
  lockTime?: string | null;
  status?: TournamentUpdateStatus;
}

export interface Golfer {
  id: string;
  tournamentId: string;
  /** @nullable */
  externalId?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  fullName: string;
  seed: number;
  /** @nullable */
  ranking?: number | null;
  /** @nullable */
  college?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  status?: string;
  eliminated: boolean;
  createdAt: string;
  updatedAt: string;
}

export type MatchupRound = typeof MatchupRound[keyof typeof MatchupRound];


export const MatchupRound = {
  R64: 'R64',
  R32: 'R32',
  R16: 'R16',
  QF: 'QF',
  SF: 'SF',
  F: 'F',
} as const;

export type MatchupStatus = typeof MatchupStatus[keyof typeof MatchupStatus];


export const MatchupStatus = {
  scheduled: 'scheduled',
  live: 'live',
  completed: 'completed',
} as const;

export interface Matchup {
  id: string;
  tournamentId: string;
  /** @nullable */
  externalId?: string | null;
  round: MatchupRound;
  roundNumber: number;
  position: number;
  /** @nullable */
  golfer1Id?: string | null;
  /** @nullable */
  golfer2Id?: string | null;
  /** @nullable */
  winnerId?: string | null;
  /** @nullable */
  nextMatchupId?: string | null;
  /** @nullable */
  nextSlot?: number | null;
  /** @nullable */
  matchScore?: string | null;
  status: MatchupStatus;
  /** @nullable */
  sourceUpdatedAt?: string | null;
  createdAt: string;
  updatedAt: string;
  golfer1?: Golfer;
  golfer2?: Golfer;
  winner?: Golfer;
}

export type MatchupUpdateStatus = typeof MatchupUpdateStatus[keyof typeof MatchupUpdateStatus];


export const MatchupUpdateStatus = {
  scheduled: 'scheduled',
  live: 'live',
  completed: 'completed',
} as const;

export interface MatchupUpdate {
  /** @nullable */
  winnerId?: string | null;
  /** @nullable */
  matchScore?: string | null;
  status?: MatchupUpdateStatus;
}

export interface BracketInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
}

export interface Bracket {
  id: string;
  userId: string;
  tournamentId: string;
  name: string;
  submitted: boolean;
  /** @nullable */
  submittedAt?: string | null;
  score: number;
  maxPossibleScore: number;
  /** @nullable */
  rank?: number | null;
  totalPicks: number;
  completedPicks: number;
  /** @nullable */
  championGolferId?: string | null;
  /** @nullable */
  championName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type BracketPickRound = typeof BracketPickRound[keyof typeof BracketPickRound];


export const BracketPickRound = {
  R64: 'R64',
  R32: 'R32',
  R16: 'R16',
  QF: 'QF',
  SF: 'SF',
  F: 'F',
} as const;

export type BracketPickStatus = typeof BracketPickStatus[keyof typeof BracketPickStatus];


export const BracketPickStatus = {
  pending: 'pending',
  correct: 'correct',
  incorrect: 'incorrect',
  eliminated: 'eliminated',
} as const;

export interface BracketPick {
  id: string;
  bracketId: string;
  matchupId: string;
  selectedGolferId: string;
  round: BracketPickRound;
  roundNumber: number;
  pointsPossible: number;
  status: BracketPickStatus;
  /** @nullable */
  isCorrect?: boolean | null;
  createdAt: string;
  updatedAt: string;
}

export interface BracketWithPicks {
  id: string;
  userId: string;
  tournamentId: string;
  name: string;
  submitted: boolean;
  /** @nullable */
  submittedAt?: string | null;
  score: number;
  maxPossibleScore: number;
  /** @nullable */
  rank?: number | null;
  totalPicks: number;
  completedPicks: number;
  /** @nullable */
  championGolferId?: string | null;
  /** @nullable */
  championName?: string | null;
  picks: BracketPick[];
  createdAt: string;
  updatedAt: string;
}

export interface PickInput {
  matchupId: string;
  selectedGolferId: string;
}

export interface LeaderboardEntry {
  rank: number;
  bracketId: string;
  bracketName: string;
  userId: string;
  userName: string;
  score: number;
  maxPossibleScore: number;
  /** @nullable */
  championName?: string | null;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
  total: number;
  /** @nullable */
  currentUserRank: number | null;
}

export interface GroupInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  description?: string;
}

export interface JoinGroupInput {
  joinCode: string;
}

export interface Group {
  id: string;
  tournamentId: string;
  name: string;
  /** @nullable */
  description?: string | null;
  joinCode: string;
  createdByUserId: string;
  memberCount: number;
  createdAt: string;
}

export interface GroupWithLeaderboard {
  id: string;
  tournamentId: string;
  name: string;
  /** @nullable */
  description?: string | null;
  joinCode: string;
  createdByUserId: string;
  memberCount: number;
  members: User[];
  leaderboard: LeaderboardEntry[];
  createdAt: string;
}

export type AdminImportInputAction = typeof AdminImportInputAction[keyof typeof AdminImportInputAction];


export const AdminImportInputAction = {
  import: 'import',
  refresh: 'refresh',
} as const;

export interface AdminImportInput {
  action: AdminImportInputAction;
}

export interface ImportResult {
  success: boolean;
  message: string;
  /** @nullable */
  golfersImported?: number | null;
  /** @nullable */
  matchupsImported?: number | null;
  /** @nullable */
  resultsUpdated?: number | null;
  errors?: string[];
  timestamp?: string;
}

export interface RescoreResult {
  success: boolean;
  bracketsRescored: number;
  message?: string;
}

export type GetLeaderboardParams = {
page?: number;
limit?: number;
};