# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bracket Golf API - U.S. Amateur Bracket Challenge
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login or register by name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in or registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user from session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /admin/import:
    post:
      operationId: adminImport
      tags: [admin]
      summary: Import or refresh USGA tournament data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /admin/tournament:
    patch:
      operationId: adminUpdateTournament
      tags: [admin]
      summary: Update tournament settings (lock time, status)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TournamentUpdate"
      responses:
        "200":
          description: Updated tournament
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
  /admin/matchups/{matchupId}:
    patch:
      operationId: adminUpdateMatchup
      tags: [admin]
      summary: Manually update matchup result
      parameters:
        - name: matchupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchupUpdate"
      responses:
        "200":
          description: Updated matchup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matchup"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
    User:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    # --- TOURNAMENT ---
    Tournament:
      type: object
      required: [id, name, year, slug, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        year:
          type: number
        slug:
          type: string
        status:
          type: string
          enum: [upcoming, active, locked, completed]
        lockTime:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        lastSyncedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TournamentUpdate:
      type: object
      properties:
        lockTime:
          type: ["string", "null"]
        status:
          type: string
          enum: [upcoming, active, locked, completed]

    # --- GOLFER ---
    Golfer:
      type: object
      required: [id, tournamentId, fullName, seed, eliminated, createdAt, updatedAt]
      properties:
        id:
          type: string
        tournamentId:
          type: string
        externalId:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        fullName:
          type: string
        seed:
          type: number
        ranking:
          type: ["integer", "null"]
        college:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        status:
          type: string
        eliminated:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string

    # --- MATCHUP ---
    Matchup:
      type: object
      required: [id, tournamentId, round, roundNumber, position, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        tournamentId:
          type: string
        externalId:
          type: ["string", "null"]
        round:
          type: string
          enum: [R64, R32, R16, QF, SF, F]
        roundNumber:
          type: number
        position:
          type: number
        golfer1Id:
          type: ["string", "null"]
        golfer2Id:
          type: ["string", "null"]
        winnerId:
          type: ["string", "null"]
        nextMatchupId:
          type: ["string", "null"]
        nextSlot:
          type: ["integer", "null"]
        matchScore:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, live, completed]
        sourceUpdatedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        golfer1:
          $ref: "#/components/schemas/Golfer"
        golfer2:
          $ref: "#/components/schemas/Golfer"
        winner:
          $ref: "#/components/schemas/Golfer"
    MatchupUpdate:
      type: object
      properties:
        winnerId:
          type: ["string", "null"]
        matchScore:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, live, completed]

    # --- BRACKET ---
...
    AdminImportInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [import, refresh]
    ImportResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        golfersImported:
          type: ["integer", "null"]
        matchupsImported:
          type: ["integer", "null"]
        resultsUpdated:
          type: ["integer", "null"]
        errors:
          type: array
          items:
            type: string
        timestamp:
          type: string
...